Fix GxEPD2_WiFi_Example bug - gotmore > 0
Why is the Issues section disabled on this git repo?

Anyway, in the Wifi example, the following lines cause
unpredictable failure loading the complete image. The
image will almost always not load completely, and part
of the top of it will be cut off. It will load fully
maybe once every 20 times, it's not working at all:

```
connection_ok = gotmore > 0;
```

This commit works around the issue by changing it to:

```
connection_ok = gotmore >= 0;
```

It seems to work perfectly, although it might not
be exactly what you wanted to be doing there. Maybe
you'd prefer to fix the logic some other way.

The issue is happening on my Waveshare 640x384 7.5"
b/w display with the GxEPD2_750 driver define
activated.

Please fix it either with this commit or however
you'd prefer, and please also enable the Issues
section on your git repo.