Update GxEPD2_750c_Z08.h
I don't know where you got the 18s for the full refresh time for this display, but it's completely wrong. WaveShare's documentation says 26s, however I have observed around 32s with the most recent ones ("V3") and up to 40s on older ones ("V2") and I remember it used to be documented as 40 or 39 seconds. I did test test it and I was getting systematically "busy timeout" with your library.